+++ common/java/com/zergatul/cheatutils/render/BlockOverlayRenderer.java
@@ -29,10 +29,10 @@ public class BlockOverlayRenderer {
     }
 
     public void quad(
-            float x1, float y1, float z1,
-            float x2, float y2, float z2,
-            float x3, float y3, float z3,
-            float x4, float y4, float z4
+            double x1, double y1, double z1,
+            double x2, double y2, double z2,
+            double x3, double y3, double z3,
+            double x4, double y4, double z4
     ) {
         triangle(
                 x1, y1, z1,
@@ -45,9 +45,9 @@ public class BlockOverlayRenderer {
     }
 
     public void triangle(
-            float x1, float y1, float z1,
-            float x2, float y2, float z2,
-            float x3, float y3, float z3
+            double x1, double y1, double z1,
+            double x2, double y2, double z2,
+            double x3, double y3, double z3
     ) {
         bufferProgram.buffer.add((float) (x1 - view.x));
         bufferProgram.buffer.add((float) (y1 - view.y));
